+++ rpc_old/rpc.go
@@ -1,8 +1,0 @@
-package rpc
-
-const (
-	packRegister = iota
-	heartBeat
-	heartBeatRet
-	packData
-)